% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiling.R
\name{aggregate_blocks}
\alias{aggregate_blocks}
\title{Aggregate raster blocks}
\usage{
aggregate_blocks(blocks_df, max_size = 1024, direction = "row", debug = FALSE)
}
\arguments{
\item{blocks_df}{A data frame of block definitions with columns nXOff, nYOff, nXSize, nYSize}

\item{max_size}{Maximum size in pixels for aggregated blocks (default 1024)}

\item{direction}{Direction of aggregation ("row" or "both")}
}
\value{
A data frame of aggregated blocks
}
\description{
Aggregate raster blocks
}
\examples{
blocks_df <- get_tiles(1000, 1000, 128, 128)
# Aggregate blocks in rows
aggregate_blocks(blocks_df, max_size = 512)
}
