% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-utils.R
\name{bbox_to_wkt}
\alias{bbox_to_wkt}
\title{Convert a bounding box to a WKT string}
\usage{
bbox_to_wkt(bbox)
}
\arguments{
\item{bbox}{A numeric vector of length 4 representing a bounding box}
}
\value{
A WKT string
}
\description{
Convert a bounding box to a WKT string
}
\keyword{internal}
