% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-block.R
\name{build_vrt_block}
\alias{build_vrt_block}
\title{constructor for vrt_block class}
\usage{
build_vrt_block(x, maskfun = NULL, pixfun = NULL, warped = FALSE, ...)
}
\arguments{
\item{x}{An xml_document object}

\item{maskfun}{A function of the mask function}

\item{pixfun}{A function of the pixel function}

\item{warped}{A logical indicating whether the VRT is warped}

\item{...}{Additional arguments not used}
}
\description{
constructor for vrt_block class
}
\keyword{internal}
