% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-to-MEM.R
\name{r_to_MEM}
\alias{r_to_MEM}
\title{Create an in-memory GDAL raster dataset from R data}
\usage{
r_to_MEM(data, bbox, srs, data_type, no_data_value, nbands, xsize, ysize)
}
\arguments{
\item{data}{numeric vector, matrix or array}

\item{bbox}{numeric vector of length 4 representing the bounding box with
the order: xmin, ymin, xmax, ymax}

\item{srs}{character spatial reference system string}

\item{data_type}{character GDAL data type name (e.g., "Float32", "Int16"),
if NULL (default) and the data has no \code{gis} attribute, "Float32" is used.}

\item{no_data_value}{numeric NoData value to set for all bands. If NULL (default)
a sensible default NoData value for the specified \code{data_type} is used.}

\item{nbands}{numeric number of bands}

\item{xsize}{numeric number of pixels in x direction}

\item{ysize}{numeric number of pixels in y direction}
}
\value{
GDALRasterDataset object
}
\description{
Create an in-memory GDAL raster dataset from R data
}
\details{
If \code{data} has been generated from \code{\link[gdalraster]{read_ds}} it will
include a \code{gis} attribute containing the necessary spatial metadata. If this
is the case, the spatial parameters are automatically extracted from the
\code{gis} attribute.
}
\examples{
v2m <- r_to_MEM(
  as.vector(c(
    sin(
      sqrt(outer(1:120 - 60, 1:110 - 55, function(x, y) x^2 + y^2)) * 0.9
    )
  )),
  nbands = 1,
  xsize = 120,
  ysize = 110
)

v2m
plot(v2m)
}
