% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-mgmt.R
\name{is_temp_dir}
\alias{is_temp_dir}
\title{Check if a path is in the temporary directory}
\usage{
is_temp_dir(path)
}
\arguments{
\item{path}{A character string of the path to check}
}
\value{
A logical indicating if the path is in the temporary directory
}
\description{
Check if a path is in the temporary directory
}
