% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp-pixel-funs.R
\name{median_numba}
\alias{median_numba}
\title{python pixel function for the median using numba}
\usage{
median_numba()
}
\value{
character of the python function
}
\description{
python pixel function for the median using numba
}
