% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp-pixel-funs.R
\name{median_numpy}
\alias{median_numpy}
\title{python pixel function for the median using numpy}
\usage{
median_numpy()
}
\value{
character of the python function
}
\description{
python pixel function for the median using numpy
}
