% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel-funs.R
\name{numba_median}
\alias{numba_median}
\title{python pixel function for the median using numba}
\usage{
numba_median()
}
\value{
character of the python function
}
\description{
python pixel function for the median using numba
}
