% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel-funs.R
\name{numpy_median}
\alias{numpy_median}
\title{python pixel function for the median using numpy}
\usage{
numpy_median()
}
\value{
character of the python function
}
\description{
python pixel function for the median using numpy
}
