% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-raster-src.R, R/plot_raster2.R
\name{plot_raster_src}
\alias{plot_raster_src}
\alias{plot.vrt_block}
\alias{plot.vrt_stack}
\alias{plot.vrt_stack_warped}
\alias{plot.vrt_collection}
\alias{plot.Rcpp_GDALRaster}
\title{plot a raster file or \code{vrt_x} object}
\usage{
plot_raster_src(
  x,
  bands = 1,
  col = grDevices::hcl.colors(10, "Viridis"),
  title = c("description", "dttm", "none"),
  rgb_trans = c("linear", "gamma", "hist", "hist_all"),
  col_tbl = NULL,
  minmax_def = NULL,
  minmax_pct_cut = NULL,
  xlim = NULL,
  ylim = NULL,
  interpolate = TRUE,
  axes = TRUE,
  main = "",
  xlab = "",
  ylab = "",
  legend = if (length(bands) == 1) TRUE else FALSE,
  digits = 2,
  na_col = "#95acbe",
  ...
)

\method{plot}{vrt_block}(x, ..., config_options = gdal_config_opts())

\method{plot}{vrt_stack}(x, ..., config_options = gdal_config_opts())

\method{plot}{vrt_stack_warped}(x, ..., config_options = gdal_config_opts())

\method{plot}{vrt_collection}(x, item, ..., config_options = gdal_config_opts())

\method{plot}{Rcpp_GDALRaster}(
  data,
  xsize = NULL,
  ysize = NULL,
  bands = NULL,
  max_pixels = 2.5e+07,
  scale_values = TRUE,
  col_tbl = NULL,
  maxColorValue = 1,
  normalize = TRUE,
  minmax_def = NULL,
  minmax_pct_cut = NULL,
  col = NULL,
  pixel_fn = NULL,
  xlim = NULL,
  ylim = NULL,
  interpolate = TRUE,
  axes = TRUE,
  main = "",
  xlab = "x",
  ylab = "y",
  legend = FALSE,
  digits = 2,
  na_col = grDevices::rgb(0, 0, 0, 0),
  mar = c(0, 0, 0, 0),
  ...
)
}
\arguments{
\item{x}{A path to a raster file}

\item{bands}{Integer vector of length 1 or 3. Band number(s) to plot.
For grayscale plots use a single band, for RGB use three bands.}

\item{col}{Colors to interpolate; must be a valid argument to
\code{\link[grDevices]{col2rgb}}. Ignored if \code{col_tbl} is provided.}

\item{title}{a character string indicating what to use as the title of the
plot. One of "description", "dttm", or "none". If "description" the band
description (name) is used, if "dttm" the datetime of the block is used, if
"none" no title is used. Ignored if main is provided.}

\item{col_tbl}{Color table as returned by \code{ds$getColorTable()}.
If provided, overrides the \code{col} parameter.}

\item{minmax_def}{Numeric vector of min/max values for color scaling.
For RGB plots, should be length 6: c(min_r, min_g, min_b, max_r, max_g, max_b).}

\item{minmax_pct_cut}{Numeric vector of length 2. Percentile cutoff values
for color scaling (e.g., c(2, 98) for 2nd and 98th percentiles).}

\item{xlim}{Numeric vector of length 2. X-axis limits in coordinate system units.}

\item{ylim}{Numeric vector of length 2. Y-axis limits in coordinate system units.}

\item{interpolate}{Logical. Whether to apply smoothing to the raster image.}

\item{axes}{Logical. Whether to draw axes and axis labels.}

\item{main}{Character string. Plot title.}

\item{xlab}{Character string. X-axis label.}

\item{ylab}{Character string. Y-axis label.}

\item{legend}{Logical. Whether to draw a color legend. Only supported for
single-band plots. Automatically detects discrete vs continuous data.}

\item{digits}{Integer. Number of decimal places for legend labels.}

\item{na_col}{Color for NA/nodata pixels. Default is transparent.}

\item{...}{Additional arguments passed to \code{\link[graphics]{plot.window}}.}

\item{config_options}{A named character vector of gdal config options to set
before attempting to read the VRT.}

\item{item}{The numeric index of the item, in the vrt_collection, to plot}

\item{data}{An object of class \code{Rcpp_GDALRaster} see
\code{\link[gdalraster]{GDALRaster}}}

\item{xsize}{Integer. Number of pixels in the x dimension for the output plot.
If \code{NULL}, uses the full raster width.}

\item{ysize}{Integer. Number of pixels in the y dimension for the output plot.
If \code{NULL}, uses the full raster height.}

\item{max_pixels}{Numeric. Maximum number of pixels to read. If
\code{xsize * ysize} exceeds this value, the raster will be downsampled.}

\item{scale_values}{Logical. Whether to apply scale and offset values
from the raster metadata.}

\item{maxColorValue}{Numeric. Maximum color value when using a color table.}

\item{normalize}{Logical. Whether to normalize pixel values to [0,1] range
before color mapping. Automatically disabled for discrete data.}

\item{pixel_fn}{Function to apply to pixel values before plotting.}

\item{mar}{Numeric vector of length 4. Additional margin adjustments
to add to the base margins when legend is enabled.}

\item{pal}{a character vector of colors to use when plotting a single band.
for example when sea presents as very dark.}
}
\description{
A wrapper around \code{gdalraster::plot_raster} that simplifies the
process of plotting a raster file.
}
\examples{
s2_imgs <- fs::dir_ls(system.file("s2-data", package = "vrtility"))
ds <- methods::new(gdalraster::GDALRaster, s2_imgs[2])
on.exit(ds$close(), add = TRUE)
plot(
  ds,
  bands = 4,
  legend = TRUE
)
}
