% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-raster-src.R
\name{plot_raster_src}
\alias{plot_raster_src}
\title{plot a raster file}
\usage{
plot_raster_src(
  x,
  bands = 1,
  pal = grDevices::hcl.colors(10, "viridis"),
  xsize = NULL,
  ysize = NULL,
  nbands = length(bands),
  max_pixels = 2.5e+07,
  col_tbl = NULL,
  maxColorValue = 1,
  normalize = TRUE,
  minmax_def = NULL,
  minmax_pct_cut = if (length(bands) == 3) c(1, 97) else NULL,
  col_map_fn = if (nbands == 1) scales::colour_ramp(pal, alpha = FALSE) else NULL,
  xlim = NULL,
  ylim = NULL,
  interpolate = TRUE,
  asp = 1,
  axes = TRUE,
  main = "",
  xlab = "x",
  ylab = "y",
  xaxs = "i",
  yaxs = "i",
  legend = if (nbands == 1) TRUE else FALSE,
  digits = 2,
  na_col = rgb(0, 0, 0, 0),
  ...
)
}
\arguments{
\item{x}{A path to a raster file}

\item{bands}{a numeric vector of band numbers to plot must be of length
1 or 3.}

\item{pal}{a character vector of colors to use when plotting a single band.}

\item{xsize}{The number of pixels along the x dimension in \code{data}. If \code{data}
is a \code{GDALRaster} object, specifies the size at which the raster will be
read (used for argument \code{out_xsize} in \code{GDALRaster$read()}). By default,
the entire raster will be read at full resolution.}

\item{ysize}{The number of pixels along the y dimension in \code{data}. If \code{data}
is a \code{GDALRaster} object, specifies the size at which the raster will be
read (used for argument \code{out_ysize} in \code{GDALRaster$read()}). By default,
the entire raster will be read at full resolution.}

\item{nbands}{The number of bands in \code{data}. Must be either 1 (grayscale) or
3 (RGB). For RGB, \code{data} are interleaved by band. If \code{nbands} is \code{NULL} (the
default), then \code{nbands = 3} is assumed if the input data contain 3 bands,
otherwise band 1 is used.}

\item{max_pixels}{The maximum number of pixels that the function will
attempt to display (per band). An error is raised if \code{(xsize * ysize)}
exceeds this value. Setting to \code{NULL} turns off this check.}

\item{col_tbl}{A color table as a matrix or data frame with four or five
columns. Column 1 contains the numeric pixel values. Columns 2:4 contain
the intensities of the red, green and blue primaries (\code{0:1} by default,
or use integer \code{0:255} by setting \code{maxColorValue = 255}).
An optional column 5 may contain alpha transparency values, \code{0} for fully
transparent to \code{1} (or \code{maxColorValue}) for opaque (the default if column 5
is missing). If \code{data} is a \code{GDALRaster} object, a built-in color table will
be used automatically if one exists in the dataset.}

\item{maxColorValue}{A number giving the maximum of the color values range
in \code{col_tbl} (see above). The default is \code{1}.}

\item{normalize}{Logical. \code{TRUE} to rescale pixel values so that their
range is \verb{[0,1]}, normalized to the full range of the pixel data by default
(\code{min(data)}, \code{max(data)}, per band). Ignored if \code{col_tbl} is used.
Set \code{normalize} to \code{FALSE} if a color map function is used that
operates on raw pixel values (see \code{col_map_fn} below).}

\item{minmax_def}{Normalize to user-defined min/max values (in terms of
the pixel data, per band). For single-band grayscale, a numeric vector of
length two containing min, max. For 3-band RGB, a numeric vector of length
six containing b1_min, b2_min, b3_min, b1_max, b2_max, b3_max.}

\item{minmax_pct_cut}{Normalize to a truncated range of the pixel data using
percentile cutoffs (removes outliers). A numeric vector of length two giving
the percentiles to use (e.g., \code{c(2, 98)}). Applied per band. Ignored if
\code{minmax_def} is used.}

\item{col_map_fn}{An optional color map function (default is
\code{grDevices::gray} for single-band data or \code{grDevices::rgb} for 3-band).
Ignored if \code{col_tbl} is used. Set \code{normalize} to \code{FALSE} if using a color
map function that operates on raw pixel values.}

\item{xlim}{Numeric vector of length two giving the x coordinate range.
If \code{data} is a \code{GDALRaster} object, the default is the raster xmin, xmax in
georeferenced coordinates, otherwise the default uses pixel/line
coordinates (\code{c(0, xsize)}).}

\item{ylim}{Numeric vector of length two giving the y coordinate range.
If \code{data} is a \code{GDALRaster} object, the default is the raster ymin, ymax in
georeferenced coordinates, otherwise the default uses pixel/line
coordinates (\code{c(ysize, 0)}).}

\item{interpolate}{Logical indicating whether to apply linear interpolation
to the image when drawing (default \code{TRUE}).}

\item{asp}{Numeric. The aspect ratio y/x (see \code{?plot.window}).}

\item{axes}{Logical. \code{TRUE} to draw axes (the default).}

\item{main}{The main title (on top).}

\item{xlab}{Title for the x axis (see \code{?title}).}

\item{ylab}{Title for the y axis (see \code{?title}).}

\item{xaxs}{The style of axis interval calculation to be used for the x axis
(see \code{?par}).}

\item{yaxs}{The style of axis interval calculation to be used for the y axis
(see \code{?par}).}

\item{legend}{Logical indicating whether to include a legend on the plot.
Currently, legends are only supported for continuous data. A color table
will be used if one is specified or the raster has a built-in color table,
otherwise the value for \code{col_map_fn} will be used.}

\item{digits}{The number of digits to display after the decimal point in
the legend labels when raster data are floating point.}

\item{na_col}{Color to use for \code{NA} as a 7- or 9-character hexadecimal code.
The default is transparent (\code{"#00000000"}, the return value of
\code{rgb(0,0,0,0)}).}

\item{...}{Other parameters to be passed to \code{plot.default()}.}
}
\description{
A wrapper around \code{gdalraster::plot_raster} that simplifies the
process of plotting a raster file.
}
