% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-funs.R
\name{print.stac_vrt}
\alias{print.stac_vrt}
\title{Print a stac_vrt object}
\usage{
\method{print}{stac_vrt}(x, xml = FALSE, ...)
}
\arguments{
\item{x}{A stac_vrt object}

\item{xml}{A logical indicating whether to print the XML}

\item{...}{Additional arguments not used}
}
\description{
Print a stac_vrt object
}
