% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomedian.R
\name{reduce_to_geomedian}
\alias{reduce_to_geomedian}
\title{Create a geometric median composite.
This function creates geometric median (aka the spatial or l1 median)
composite of a warped VRT collection.}
\usage{
reduce_to_geomedian(
  x,
  outfile = fs::file_temp(ext = "tif"),
  cache_dir = getOption("vrt.cache"),
  config_options = gdal_config_opts(),
  compression = "LZW",
  return_internals = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{x}{A vrt_collection_warped object.}

\item{outfile}{The output file path.}

\item{cache_dir}{The directory to save temporary files. (default should be
fine - this is here mainly for explicit privision for mirai daemons)}

\item{config_options}{A named character vector of GDAL configuration options.}

\item{compression}{The compression method to use for the output file.}

\item{return_internals}{Logical indicating whether to return the internals of
the function (e.g., the jobs object) that is produced before promise
evaluation.}

\item{quiet}{Logical indicating whether to suppress the progress bar.}
}
\value{
A character vector of the output file path.
}
\description{
Create a geometric median composite.
This function creates geometric median (aka the spatial or l1 median)
composite of a warped VRT collection.
}
