% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac-utils.R
\name{sentinel2_stac_collect}
\alias{sentinel2_stac_collect}
\title{Generate a Sentinel 2 stac collection doc_imes object}
\usage{
sentinel2_stac_collect(
  bbox,
  start_date,
  end_date,
  assets = c("B01", "B02", "B03", "B04", "B05", "B06", "B07", "B08", "B8A", "B09", "B11",
    "B12", "SCL"),
  max_cloud_cover = 10,
  stac_source = "https://planetarycomputer.microsoft.com/api/stac/v1/",
  collection = "sentinel-2-l2a",
  mpc_sign = TRUE
)
}
\arguments{
\item{bbox}{A numeric vector of the bounding box (length 4) in lat/long}

\item{start_date}{A character string of the start date}

\item{end_date}{A character string of the end date}

\item{assets}{A character vector of the asset names to include}

\item{max_cloud_cover}{A numeric value of the maximum cloud cover percentage}

\item{stac_source}{A character string of the STAC source}

\item{collection}{A character string of the collection to query}
}
\value{
A stac_vrt object
}
\description{
Generate a Sentinel 2 stac collection doc_imes object
}
