% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-tools.R
\name{set_vrt_descriptions}
\alias{set_vrt_descriptions}
\title{Internal function to set the descriptions for the bands in a VRT}
\usage{
set_vrt_descriptions(x, descriptions, as_file = FALSE)
}
\arguments{
\item{x}{A character string of the path to the VRT}

\item{descriptions}{A character vector of the descriptions}

\item{as_file}{A logical indicating if the VRT should be saved to a file}
}
\value{
A character string of the path to the modified VRT
}
\description{
Internal function to set the descriptions for the bands in a VRT
}
\keyword{internal}
