% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-tools.R
\name{set_vrt_metadata}
\alias{set_vrt_metadata}
\title{Internal function to set metadata in a VRT}
\usage{
set_vrt_metadata(x, keys, values, as_file = FALSE)
}
\arguments{
\item{x}{A character string of the path to the VRT}

\item{keys}{A character vector of the metadata keys}

\item{values}{A character vector of the metadata values}

\item{as_file}{A logical indicating if the VRT should be saved to a file}
}
\value{
An xml_document object
}
\description{
Internal function to set metadata in a VRT
}
\keyword{internal}
