% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac-utils.R
\name{sign_planetary_computer}
\alias{sign_planetary_computer}
\title{Sign STAC items retrieved from the Planetary Computer}
\usage{
sign_planetary_computer(items, subscription_key = Sys.getenv("rsi_pc_key"))
}
\arguments{
\item{items}{A STACItemCollection.}

\item{subscription_key}{Optionally, a subscription key associated with your
Planetary Computer account. At the time of writing, this is required for
downloading Sentinel 1 RTC products, as well as NAIP imagery. This key will
be automatically used if the environment variable \code{rsi_pc_key} is set.}
}
\value{
A STACItemCollection object with signed assets url.
}
\description{
Sign STAC items retrieved from the Planetary Computer
}
\details{
copied from the \code{rsi} package
}
