% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac-utils.R
\name{stac_query}
\alias{stac_query}
\title{Query a STAC source}
\usage{
stac_query(
  bbox,
  stac_source,
  collection,
  start_date,
  end_date,
  limit = 999,
  ...
)
}
\arguments{
\item{bbox}{A numeric vector of length 4 representing a bounding box}

\item{stac_source}{The STAC source to query}

\item{collection}{The collection to query}

\item{start_date}{The start date for the query}

\item{end_date}{The end date for the query}

\item{limit}{The number of items to return}

\item{...}{Additional arguments to pass to the query}
}
\value{
A list of items
}
\description{
Query a STAC source
}
