% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac-vrt-class.R
\name{print.stac_vrt}
\alias{print.stac_vrt}
\alias{save_vrt}
\alias{save_vrt.stac_vrt}
\title{Print a stac_vrt object}
\usage{
\method{print}{stac_vrt}(x, xml = FALSE, pixfun = FALSE, ...)

save_vrt(x, file)

\method{save_vrt}{stac_vrt}(
  x,
  outfile = fs::file_temp(tmp_dir = getOption("vrt.cache"), ext = "vrt")
)
}
\arguments{
\item{x}{A stac_vrt object}

\item{xml}{A logical indicating whether to print the XML}

\item{...}{Additional arguments not used}
}
\description{
Print a stac_vrt object

Save a stac_vrt object to disk
}
