% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-utils.R
\name{validate_bbox}
\alias{validate_bbox}
\title{Validate a bounding box}
\usage{
validate_bbox(bbox)
}
\arguments{
\item{bbox}{A numeric vector of length 4 representing a bounding box}
}
\value{
A numeric vector of length 4 representing a bounding box
}
\description{
Validate a bounding box
}
\examples{
validate_bbox(c(-180, -90, 180, 90))
}
