% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-add-empty-band.R
\name{vrt_add_empty_band}
\alias{vrt_add_empty_band}
\title{Add an empty band to a VRT_x object}
\usage{
vrt_add_empty_band(x, after, description, scale_value)
}
\arguments{
\item{x}{A VRT_x object}

\item{after}{numeric indicating the band number to add the empty band after}

\item{description}{A character string describing the empty band}

\item{scale_value}{A numeric value to set the scale of the new band. If NULL,
the scale of the first band in the VRT will be used. Be careful, Landsat for
example has different scales for different bands.}
}
\description{
Add an empty band to a VRT_x object
}
