% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-funs.R
\name{vrt_add_pixfun}
\alias{vrt_add_pixfun}
\title{Compose a standard VRT for compositing at the band level}
\usage{
vrt_add_pixfun(src, vrt_opts = NULL, pix_fun = vrtility::numba_median)
}
\arguments{
\item{vrt_opts}{A list of options to pass to the VRT}

\item{pix_fun}{A character string of the pixel function code}

\item{src_files}{A character vector of file paths to the source rasters}
}
\value{
A character string of the path to the modified VRT
}
\description{
Compose a standard VRT for compositing at the band level
}
\details{
This function will add a pixel function to every band in a VRT file. Then,
when passed to gdal warp or translate the pixel function will be applied.
}
