% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac-vrt-class.R
\name{print.vrt_stack}
\alias{print.vrt_stack}
\alias{print.vrt_collection}
\alias{save_vrt}
\alias{save_vrt.vrt_stack}
\title{Print a vrt_block object}
\usage{
\method{print}{vrt_stack}(x, xml = FALSE, pixfun = FALSE, ...)

\method{print}{vrt_collection}(x, xml = FALSE, pixfun = FALSE, blocks = FALSE, ...)

save_vrt(x, file)

\method{save_vrt}{vrt_stack}(
  x,
  outfile = fs::file_temp(tmp_dir = getOption("vrt.cache"), ext = "vrt")
)
}
\arguments{
\item{x}{A vrt_block object}

\item{xml}{A logical indicating whether to print the XML}

\item{...}{Additional arguments not used}
}
\description{
Print a vrt_block object

Save a vrt_block object to disk
}
