% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-save.R
\name{vrt_block_save_internal}
\alias{vrt_block_save_internal}
\title{Internal function for saving a vrt block overcoming issues around parallel
process temp directories.}
\usage{
vrt_block_save_internal(x, temp_vrt_dir, apply_scale = TRUE)
}
\description{
Internal function for saving a vrt block overcoming issues around parallel
process temp directories.
}
\keyword{internal}
