% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-mgmt.R
\name{vrt_cache_set}
\alias{vrt_cache_set}
\alias{vrt_cache_destroy}
\title{Set the VRT cache directory}
\usage{
vrt_cache_set(dir)

vrt_cache_destroy()
}
\arguments{
\item{dir}{A character string of the directory to set as the cache}
}
\value{
A character string of the directory (invisibly)
}
\description{
Set the VRT cache directory

Destroy the VRT cache directory
}
\details{
The default cache location is \code{tempdir()}. This function allows you to set
the cache location to a different directory. It is mainly useful if you intend
to work with parallel processing with, for example {mirai} or {future}.
{vrtility} depends on saving many vrt files to disk and passing these files
between processes fails when using a temporary directory.

A helper function that will destroy the VRT cache directory. This is useful
if you want to clear the cache directory and start fresh. If the cache
directory is set to \code{tempdir()} then the function will warn you that this is
a bad idea and will not proceed.
}
