% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-classes.R, R/vrt-collect.R, R/vrt-stack.R
\name{print.vrt_stack}
\alias{print.vrt_stack}
\alias{print.vrt_collection}
\alias{save_vrt}
\alias{save_vrt.vrt_stack}
\alias{vrt_collect}
\alias{vrt_collect.doc_items}
\alias{vrt_stack}
\alias{vrt_stack.vrt_collection}
\title{Print a vrt_block object}
\usage{
\method{print}{vrt_stack}(x, xml = FALSE, pixfun = FALSE, maskfun = FALSE, ...)

\method{print}{vrt_collection}(x, xml = FALSE, pixfun = FALSE, maskfun = FALSE, blocks = FALSE, ...)

save_vrt(x, outfile)

\method{save_vrt}{vrt_stack}(
  x,
  outfile = fs::file_temp(tmp_dir = getOption("vrt.cache"), ext = "vrt")
)

vrt_collect(x, t_srs, te, tr, quiet = FALSE)

\method{vrt_collect}{doc_items}(x, t_srs, te, tr, quiet = FALSE)

vrt_stack(x, ...)

\method{vrt_stack}{vrt_collection}(x, quiet = TRUE, ...)
}
\arguments{
\item{x}{An object to be used to create a vrt_x object see details.}

\item{xml}{A logical indicating whether to print the XML}

\item{pixfun}{A logical indicating whether to print the pixel function}

\item{maskfun}{A logical indicating whether to print the mask function}

\item{...}{Additional arguments passed to methods}

\item{blocks}{A logical indicating whether to print the blocks instead of
the collection summary.}

\item{outfile}{A character string of the output file}

\item{t_srs}{character target SRS must be a numeric EPSG code, or SRS like
character such as a proj4 string or WKT.}

\item{te}{numeric vector of the target extent in the form
c(xmin, ymin, xmax, ymax) using the same SRS as in \code{t_srs}.}

\item{tr}{numeric vector of the target resolution in the form c(xres, yres)}

\item{quiet}{Logical. If TRUE, suppress GDAL progress bar}
}
\description{
Print a vrt_block object

Save a vrt_block object to disk

Construct the base VRT object for composing VRT pipelines.

Stack VRT files from a vrt_collection object
}
\details{
For now the main way to create a vrt_collection object, which forms
the basis of the vrrt-based pipelines in vrtility is using a doc_items
object from the \code{rstac} package. For more info on how to create a doc_items
object see \code{\link[=sentinel2_stac_query]{sentinel2_stac_query()}}. To build a vrt_stack object a vrt_collection
is required first. The vrt_collection object is essentially a list of warped
VRT files. When vrt_stack is called on a vrt_collection object, the VRT files
are stacked into a single VRT file where each band in the VRT contains the
files of multiple epochs. We can then composite or summarise these layers
using the pixel functions.
}
