% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-collect.R
\name{vrt_collect}
\alias{vrt_collect}
\alias{vrt_collect.doc_items}
\title{Construct the base VRT object for composing VRT pipelines.}
\usage{
vrt_collect(x)

\method{vrt_collect}{doc_items}(x)
}
\description{
Construct the base VRT object for composing VRT pipelines.
}
