% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-composites.R
\name{vrt_composite}
\alias{vrt_composite}
\title{Generate a composite raster from (virtual) raster sources.}
\usage{
vrt_composite(
  src_files,
  outfile,
  bbox,
  t_srs = to_generic_projected(bbox),
  res = 10,
  warp_options = c("-overwrite", "-te", gdalraster::bbox_transform(bbox,
    gdalraster::srs_to_wkt("EPSG:4326"), to_generic_projected(bbox)), "-tr", res, res,
    "-tap", "-r", "bilinear", "-co", "COMPRESS=DEFLATE", "-co", "PREDICTOR=2", "-co",
    "NUM_THREADS=ALL_CPUS"),
  config_options = c(GDAL_VRT_ENABLE_PYTHON = "YES", VSI_CACHE = "TRUE", GDAL_CACHEMAX =
    "30\%", VSI_CACHE_SIZE = "10000000", GDAL_HTTP_MULTIPLEX = "YES",
    GDAL_INGESTED_BYTES_AT_OPEN = "32000", GDAL_DISABLE_READDIR_ON_OPEN = "EMPTY_DIR",
    GDAL_HTTP_VERSION = "2", GDAL_HTTP_MERGE_CONSECUTIVE_RANGES = "YES", GDAL_NUM_THREADS
    = "ALL_CPUS"),
  quiet = FALSE
)
}
\arguments{
\item{src_files}{A stac_vrt object or a character vector of file paths to the
rasters}

\item{outfile}{A character string of the output file path}

\item{bbox}{A numeric vector of the bounding box (length 4) in lat/long}

\item{t_srs}{A character string of the target SRS}

\item{res}{A numeric of the target resolution in units of t_srs}

\item{warp_options}{A character vector of options to pass to the warp}

\item{config_options}{A character vector of options to set in the GDAL
environment}

\item{quiet}{A logical indicating whether to suppress output}

\item{fun}{A character string of the pixel function to apply (only median is
supported for now...)}

\item{vrt_options}{A list of options to pass to the VRT}
}
\value{
A character string of the path to the output raster
}
\description{
Generate a composite raster from (virtual) raster sources.
}
\details{
This is the main package function at present. It is quite targetted
towards Sentinel 2 data and will likely change - but this will server as the
main testing point for now!
}
