% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-create-mask.R, R/zvrt-pixel-funs-mask.R
\name{vrt_create_mask}
\alias{vrt_create_mask}
\alias{create_omnicloudmask}
\title{Create a new mask for a vrt object
This function allows you to create a mask for a VRT object based on specified input bands and a mask function.}
\usage{
vrt_create_mask(x, inbands, maskfun, nodata_value = 0)

create_omnicloudmask(
  patch_size = 1000,
  patch_overlap = 300,
  batch_size = 1,
  inference_dtype = c("bfloat16", "float32"),
  inference_device = NULL,
  nodata_value = 0,
  model_version = NULL
)
}
\arguments{
\item{x}{A VRT_x object.}

\item{inbands}{A named numeric vector of input band indices to be used for
the mask.}

\item{maskfun}{A character containing a python pixel function to be applied
to the input bands to create the mask. Currently the only provided function
is \code{\link{create_omnicloudmask}}. This string must also have three key
attributes: \code{mask_name}, \code{mask_description}, and \code{required_bands}.}

\item{nodata_value}{The nodata value to use in the output mask (numeric,
default: 0).}

\item{patch_size}{The size/dimension of the patches to use for prediction
(numeric default: 600).}

\item{patch_overlap}{The overlap between patches (numeric default: 300).}

\item{batch_size}{The batch size to use for prediction (numeric default: 1).}

\item{inference_dtype}{The data type to use for inference. Options include
"bfloat16" and "float32" (character, default: "bfloat16"). using "bfloat16"
should be faster if supported by the hardware.}

\item{inference_device}{The device to use for inference. If \code{NULL}, the
function will automatically select the best available device (character,
default: \code{NULL}). Options include "cpu", "cuda", "mps", etc. The order of
selection is based on availability: "cuda" > "mps" > "cpu".}

\item{model_version}{The version of the OmniCloudMask model to use; options
include "4.0", "3.0", "2.0", and "1.0". If \code{NULL}, the latest version will be
used (character, default: \code{NULL}).}
}
\value{
A VRT_x object with the the new mask band added.

A Python function that can be used as a pixel function in a VRT
raster. The function will apply the OmniCloudMask model to the specified
bands and create a cloud mask.
}
\description{
This function constructs a cloud mask using the OmniCloudMask
python library. It is designed to be used with the \code{vrt_set_maskfun()}
function.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
 s2files <- fs::dir_ls(system.file("s2-data", package = "vrtility"))
 ex_collect <- vrt_collect(s2files[4:5])

 ex_mask <- vrt_create_mask(
   ex_collect,
   c(red = 3, green = 2, nir = 4),
   maskfun = create_omnicloudmask()
 )

 ex_mask_compute <- vrt_compute(
   ex_mask,
   fs::file_temp(ext = ".tif"),
   recollect = TRUE
 )
 plot(ex_mask_compute, item = 2, bands =   6)
\dontshow{\}) # examplesIf}
}
\references{
OmniCloudMask GitHub repository: \url{https://github.com/DPIRD-DMA/OmniCloudMask}
}
