% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-derived-block.R
\name{vrt_derived_block}
\alias{vrt_derived_block}
\title{Create vrt blocks with derived bands.}
\usage{
vrt_derived_block(x, ...)
}
\arguments{
\item{...}{named R formulas for new bands to create. Formula variable names
must the names of existing bands in the vrt_block object. the argument name
is used to name the new (derived) band.}
}
\description{
Create new vrt blocks containing derived expression bands.
}
