% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-move-band.R
\name{vrt_move_band}
\alias{vrt_move_band}
\title{Move a band in a VRT_x object}
\usage{
vrt_move_band(x, band_idx, after)
}
\arguments{
\item{x}{A VRT_x object}

\item{band_idx}{numeric indicating the band number of the band to move.}

\item{after}{numeric indicating the band after which the moved band should be
placed. Note this is based on the initial state of the band ordering. eg. do
not add 1 if you are moving the band forward.}
}
\value{
A modified object of the same class as \code{x} with the band reordered.
}
\description{
Move a band in a VRT_x object
}
