% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-funs.R
\name{vrt_set_band_names}
\alias{vrt_set_band_names}
\title{set the band names for a vrt_x object}
\usage{
vrt_set_band_names(x, band_names)
}
\arguments{
\item{x}{A vrt_block, vrt_stack, or vrt_collection object}

\item{band_names}{A character vector of the band names}
}
\description{
set the band names for a vrt_x object
}
