% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-set-gdal-pixfun.R,
%   R/zvrt-pixel-funs-composite-cpp.R
\name{vrt_set_gdal_pixelfun}
\alias{vrt_set_gdal_pixelfun}
\alias{vrt_set_gdal_pixelfun.vrt_block}
\alias{vrt_set_gdal_pixelfun.vrt_collection}
\alias{median_cpp}
\alias{mean_cpp}
\alias{geomean_cpp}
\alias{quantile_cpp}
\alias{mean_db_cpp}
\title{Set built-in GDAL pixel functions of a VRT stack object}
\usage{
vrt_set_gdal_pixelfun(x, pixfun, ..., band_idx, nodata_as_nan)

\method{vrt_set_gdal_pixelfun}{vrt_block}(x, pixfun, ..., band_idx = NULL, nodata_as_nan = FALSE)

\method{vrt_set_gdal_pixelfun}{vrt_collection}(x, pixfun, ..., band_idx = NULL, nodata_as_nan = FALSE)

median_cpp()

mean_cpp()

geomean_cpp()

quantile_cpp(q = 0.5)

mean_db_cpp()
}
\arguments{
\item{x}{A vrt_stack object}

\item{pixfun}{A built-in GDAL pixel function. See details for a list of
available functions.}

\item{...}{named arguments used within the pixel function see details.}

\item{band_idx}{The indices of the bands to set the pixel function for. If
NULL, the pixel function is set for all bands.}

\item{nodata_as_nan}{A logical indicating if the NoData value should be set to
NaN. Defaults to FALSE. Sometimes useful if the pixel functions ignore the
NoData value. see https://github.com/OSGeo/gdal/issues/4746. If you use this
you will probably need to specify the '-srcnodata' option on compute.}

\item{q}{Probability of the quantile to compute. Values must be between
0 and 1.}
}
\value{
A modified object of the same class as \code{x} with the GDAL pixel
function set.
}
\description{
Set built-in GDAL pixel functions of a VRT stack object

These functions return the name of a registered C++ pixel
function for use with \code{vrt_set_gdal_pixelfun()}. They are faster than
the Python equivalents and don't require Python dependencies.
Requires GDAL >= 3.4 at compile time.
}
\details{
The documentation for the GDAL pixel functions can be found at
\url{https://gdal.org/en/stable/drivers/raster/vrt.html#built-in-pixel-functions}

Where a pixel function requires arguments, these can be passed as named
arguments, in line with the specified naming in PixelFunctionArguments of
the table at the above link.

\code{median_cpp} is a C++ pixel function that calculates the median of
the input arrays, skipping NoData values.

\code{mean_cpp} is a C++ pixel function that calculates the mean of the
input arrays, skipping NoData values.

\code{geomean_cpp} is a C++ pixel function that calculates the geometric
mean of the input arrays, skipping NoData values.

\code{quantile_cpp} is a C++ pixel function that calculates the quantile
of the input arrays, skipping NoData values.

\code{mean_db_cpp} is a C++ pixel function that calculates the mean of
the input arrays in decibels, skipping NoData values. Useful for radar
backscatter data.
}
\examples{
s2files <- fs::dir_ls(system.file("s2-data", package = "vrtility"))
ex_collect <- vrt_collect(s2files)

t_block <- ex_collect[[1]][[1]]
ex_stack <- ex_collect |>
  vrt_set_maskfun(
    mask_band = "SCL",
    mask_values = c(0, 1, 2, 3, 8, 9, 10, 11),
    drop_mask_band = TRUE
  ) |>
  vrt_warp(
    t_srs = t_block$srs,
    te = t_block$bbox,
    tr = t_block$res
  ) |>
  vrt_stack()

plot(
  vrt_set_gdal_pixelfun(
    ex_stack,
    pixfun = "min",
    propagateNoData = TRUE
  ),
  bands = c(3, 2, 1)
)

plot(
  vrt_set_gdal_pixelfun(
    ex_stack,
    pixfun = "min"
  ),
  bands = c(3, 2, 1)
)

}
