% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-set-gdal-pixfun.R
\name{vrt_set_gdal_pixfun}
\alias{vrt_set_gdal_pixfun}
\alias{vrt_set_gdal_pixfun.vrt_stack}
\title{Set built-in GDAL pixel functions of a VRT stack object}
\usage{
vrt_set_gdal_pixfun(x, pixfun, ..., band_idx)

\method{vrt_set_gdal_pixfun}{vrt_stack}(x, pixfun, ..., band_idx = NULL)
}
\arguments{
\item{x}{A vrt_stack object}

\item{pixfun}{A built-in GDAL pixel function. See details for a list of
available functions.}

\item{...}{named arguments used within the pixel function see details.}

\item{band_idx}{The indices of the bands to set the pixel function for. If
NULL, the pixel function is set for all bands.}
}
\description{
Set built-in GDAL pixel functions of a VRT stack object
}
\details{
The documentation for the GDAL pixel functions can be found at
\url{https://gdal.org/en/stable/drivers/raster/vrt.html#built-in-pixel-functions}
}
