% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask-pixel-funs.R, R/vrt_set_mask.R
\name{bitmask_numpy}
\alias{bitmask_numpy}
\alias{bitmask_numba}
\alias{vrt_set_mask}
\alias{vrt_set_mask.vrt_block}
\alias{vrt_set_mask.vrt_collection}
\title{Masking functions for pixel-based processing}
\usage{
bitmask_numpy()

bitmask_numba()

vrt_set_mask(x, ...)

\method{vrt_set_mask}{vrt_block}(x, mask_band, valid_bits, mask_pix_fun = vrtility::bitmask_numba)

\method{vrt_set_mask}{vrt_collection}(x, mask_band, valid_bits, mask_pix_fun = vrtility::bitmask_numba)
}
\arguments{
\item{x}{A VRT collection}

\item{mask_band}{The name of the mask band}

\item{valid_bits}{A numeric vector of valid bits}

\item{mask_pix_fun}{A function that returns the Python code for the mask
pixel function}
}
\value{
A VRT block with the mask band set.

A VRT collection with the mask band set.
}
\description{
Masking functions for pixel-based processing

Masking functions for pixel-based processing

Set mask band of a VRT collection
}
\details{
This function is a simple bit mask function that can be used to
mask out pixels based on a bit mask. There is no bitwise transformation
applied to the provided valid_bits.

This function is a simple bit mask function that can be used to
mask out pixels based on a bit mask. There is no bitwise transformation
applied to the provided valid_bits.
}
