% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-set-mask.R, R/zvrt-pixel-funs-mask.R
\name{vrt_set_maskfun}
\alias{vrt_set_maskfun}
\alias{vrt_set_maskfun.vrt_block}
\alias{vrt_set_maskfun.vrt_collection}
\alias{build_intmask}
\alias{build_bitmask}
\title{Set mask band of a VRT collection}
\usage{
vrt_set_maskfun(
  x,
  mask_band,
  mask_values,
  build_mask_pixfun = NULL,
  buffer_size = 0,
  drop_mask_band = TRUE
)

\method{vrt_set_maskfun}{vrt_block}(
  x,
  mask_band,
  mask_values,
  build_mask_pixfun = NULL,
  buffer_size = 0,
  drop_mask_band = TRUE
)

\method{vrt_set_maskfun}{vrt_collection}(
  x,
  mask_band,
  mask_values,
  build_mask_pixfun = NULL,
  buffer_size = 0,
  drop_mask_band = TRUE
)

build_intmask()

build_bitmask()
}
\arguments{
\item{x}{A VRT collection}

\item{mask_band}{The name of the mask band}

\item{mask_values}{A numeric vector of integer or bit values to be masked.}

\item{build_mask_pixfun}{A C++ pixel function for building the mask. If
\code{NULL} (default), uses \code{\link[=build_intmask]{build_intmask()}} which returns a registered C++
pixel function. Provided functions include \code{\link[=build_intmask]{build_intmask()}} and
\code{\link[=build_bitmask]{build_bitmask()}}. Python pixel functions are only supported when
\code{buffer_size > 0}.}

\item{buffer_size}{A buffer size to apply to the mask (numeric, default: 0). A buffer
size > 0 will dilate the mask by the specified number of pixels.
This can be useful to remove edge effects around clouds.
If a buffer size > 0 is specified, the \code{scipy} python library will
automatically be installed and the legacy Python masking path will be used.}

\item{drop_mask_band}{Logical. If TRUE, the mask band will be removed from
the VRT block.}
}
\value{
A VRT block with the mask band set.

A VRT collection with the mask band set.
}
\description{
Set mask band of a VRT collection
}
\details{
Masking uses GDAL's native VRT MaskBand mechanism (RFC 15). A C++ pixel
function converts classification values to a standard mask (0=invalid,
255=valid), which is set as a dataset-level MaskBand. An outer VRT then
references the inner VRT using ComplexSource with UseMaskBand=true, so
GDAL natively handles the masking.

For integer-based masking (e.g. Sentinel-2 SCL band), use \code{\link[=build_intmask]{build_intmask()}}.
For bitwise masking (e.g. HLS Fmask), use \code{\link[=build_bitmask]{build_bitmask()}}.

\code{build_intmask} provides an integer mask function that can be used
to mask out pixels based on a band containing true integer/numeric values.
This would be appropriate for the Sentinel 2A SCL band, for example.

\code{build_bitmask} provides a simple bit-wise mask function that can
be used to mask out pixels based on a true bit mask. This function should be
used where bitwise operations are required. e.g. for HLS data, the "Fmask"
band requires bitwise operations to extract the mask values.
}
\examples{
s2files <- fs::dir_ls(system.file("s2-data", package = "vrtility"))

ex_collect <- vrt_collect(s2files)

ex_collect |>
  vrt_set_maskfun(
    mask_band = "SCL",
    mask_values = c(0, 1, 2, 3, 8, 9, 10, 11),
    drop_mask_band = FALSE)

}
