% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-set-nodata.R
\name{vrt_set_nodata}
\alias{vrt_set_nodata}
\alias{vrt_set_nodata.vrt_block}
\alias{vrt_set_nodata.vrt_collection}
\title{Set NoData Value for VRT}
\usage{
vrt_set_nodata(x, nodatavalue, nodata, band_idx)

\method{vrt_set_nodata}{vrt_block}(x, nodatavalue, nodata = nodatavalue, band_idx = NULL)

\method{vrt_set_nodata}{vrt_collection}(x, nodatavalue, nodata = nodatavalue, band_idx = NULL)
}
\arguments{
\item{x}{A VRT_x object}

\item{nodatavalue}{A numeric value to set the NoDataValue XML tag for the VRT_x
object.}

\item{nodata}{A numeric value to set the NODATA XML tag for the VRT_x
object. defaults to the same as \code{nodatavalue}.}

\item{band_idx}{numeric; the target band position(s) to set the NoData value
for. If NULL, The nodatavalue will be set for all bands.}
}
\value{
A modified object of the same class as \code{x} with the NoData value set.
}
\description{
Set the NoData value for a VRT_x object. This information is
preserved in derived VRT and GeoTIFF files.
}
\examples{
s2files <- fs::dir_ls(system.file("s2-data", package = "vrtility"))
ex_collect <- vrt_collect(s2files)
ex_nodata <- vrt_set_nodata(
  ex_collect,
  nodatavalue = 0,
  nodata = 0
)
print(ex_nodata[[1]][[1]], xml = TRUE)
}
