% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp-pixel-funs.R, R/vrt-set-pixfun.R
\name{median_numpy}
\alias{median_numpy}
\alias{mean_numpy}
\alias{vrt_set_pixelfun}
\alias{vrt_set_pixelfun.vrt_stack}
\title{python pixel function for the mean using numpy}
\usage{
median_numpy()

mean_numpy()

vrt_set_pixelfun(x, pixfun)

\method{vrt_set_pixelfun}{vrt_stack}(x, pixfun = vrtility::median_numpy())
}
\arguments{
\item{x}{A vrt_stack object}

\item{pixfun}{A function that returns the Python code for the pixel function}
}
\value{
character of the python function

character of the python function
}
\description{
python pixel function for the mean using numpy

Set the pixel function of a VRT stack object
}
\details{
\code{median_numpy} is a pixel function that calculates the median of
the input arrays, it is injected into the VRT file as a Python function.
\code{mean_numpy} works in the same way but calculates the mean.`
}
