% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-set-py-pixfun.R,
%   R/zvrt-pixel-funs-composite.R
\name{vrt_set_py_pixelfun}
\alias{vrt_set_py_pixelfun}
\alias{vrt_set_py_pixelfun.vrt_block}
\alias{vrt_set_py_pixelfun.vrt_collection}
\alias{median_numpy}
\alias{mean_numpy}
\alias{geomean_numpy}
\alias{quantile_numpy}
\alias{mean_db_numpy}
\title{Set the pixel function of a VRT stack object}
\usage{
vrt_set_py_pixelfun(x, pixfun, band_idx)

\method{vrt_set_py_pixelfun}{vrt_block}(x, pixfun = vrtility::median_numpy(), band_idx = NULL)

\method{vrt_set_py_pixelfun}{vrt_collection}(x, pixfun = vrtility::median_numpy(), band_idx = NULL)

median_numpy()

mean_numpy()

geomean_numpy()

quantile_numpy(q, use_fastnanquantile = TRUE)

mean_db_numpy()
}
\arguments{
\item{x}{A vrt_stack object}

\item{pixfun}{A function that returns the Python code for the pixel function}

\item{band_idx}{The indices of the bands to set the pixel function for. If
NULL, the pixel function is set for all bands.}

\item{q}{Probability of the quantile to compute. Values must be between 0 and 1
inclusive.}

\item{use_fastnanquantile}{Logical indicating whether to use the
\code{fastnanquantile} library for calculating the quantile. This is generally
faster than the standard numpy implementation, especially for large arrays.
However, it requires the \code{fastnanquantile} library to be installed in the
Python environment (This is handled automatically). Default is TRUE.}
}
\value{
A modified object of the same class as \code{x} with the Python pixel
function set.

character of the python function

character of the python function
}
\description{
Set the pixel function of a VRT stack object
}
\details{
\code{median_numpy} is a pixel function that calculates the median of
the input arrays, it is injected into the VRT file as a Python function.
\code{mean_numpy} works in the same way but calculates the mean.`

\code{geomean_numpy} is a pixel function that calculates the geometric
mean of the input arrays. Use cases of this are at present unclear to me.
If you have thoughts or references please let me know.

\code{quantile_numpy} is a pixel function that calculates the quantile
of the input arrays for a given probability. This could be useful where
the median fails to filter cloudy pixels effectively. The defauly numpy
nanquantile function is very slow and does not support masked arrays.
Therefore, it is typically much faster to use the fastnanquantile library
which uses numba to increase performance. However, this library has some
limitations around multi-threading in certain contexts. Therefore, when
using this pixel function with \code{vrt_compute()}, it is strongly advised to
use the \code{warp} engine and to avoid using mirai daemons (i.e. use
sequential processing).

\code{mean_db_numpy} is a pixel function that calculates the mean of
the input arrays and then converts to decibels. This is useful for
calculating the mean of radar raw/linear backscatter values, for example.
}
