% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-stack.R
\name{vrt_stack}
\alias{vrt_stack}
\alias{vrt_stack.vrt_collection}
\title{Stack VRT files from a vrt_collection object}
\usage{
vrt_stack(x, ...)

\method{vrt_stack}{vrt_collection}(x, ...)
}
\arguments{
\item{x}{A vrt_collection object}

\item{...}{Additional arguments passed to methods}
}
\description{
Stack VRT files from a vrt_collection object
}
