% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/vrt-tools.R
\docType{data}
\name{vrt_xml_schema}
\alias{vrt_xml_schema}
\alias{vrt_schema}
\title{The official GDAL VRT XML schema
A copy of the official VRT schema for vritility VRT validation.}
\format{
character
}
\source{
https://raw.githubusercontent.com/OSGeo/gdal/master/frmts/vrt/data/gdalvrt.xsd
}
\usage{
vrt_xml_schema

vrt_schema()
}
\value{
An xml_document object
}
\description{
Loads a copy of the official xml schema from the provided
vrt_xml_schema dataset.
}
\examples{
vrt_schema()
}
\keyword{datasets}
