% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-utils.R, R/vrt-set-mask.R
\name{to_wkt}
\alias{to_wkt}
\alias{to_wkt.default}
\alias{to_wkt.numeric}
\alias{to_wkt.character}
\alias{vrt_set_maskfun.vrt_block}
\title{Convert an object to wkt.}
\usage{
to_wkt(x)

\method{to_wkt}{default}(x)

\method{to_wkt}{numeric}(x)

\method{to_wkt}{character}(x)

\method{vrt_set_maskfun}{vrt_block}(
  x,
  valid_bits,
  mask_band,
  mask_pix_fun = vrtility::bitmask_numba,
  drop_mask_band = TRUE
)
}
\arguments{
\item{x}{A VRT block}

\item{valid_bits}{A numeric vector of valid bits}

\item{mask_band}{The name of the mask band}

\item{mask_pix_fun}{A function that returns the Python code for the mask
pixel function}

\item{drop_mask_band}{Logical. If TRUE, the mask band will be removed from
the VRT block.}
}
\value{
A character string of the object in wkt format.

A VRT block with the mask band set.
}
\description{
Convert an object to wkt.
}
\keyword{internal}
