% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{vrtility}
\alias{vrtility}
\alias{vrtility-package}
\title{GDAL VRT utilities for R}
\description{
THE GDAL VRT (Virtual Format) is an extremely powerful tool for combining
and manipulating raster data. This package provides a set of utilities
for working with VRTs in R. It is primarily focussed on harnessing the
power of VRT pixel functions to carry out complex raster operations. This
is achieved through the use of python and in particular the numba library
which allows for the creation of fast, compiled functions that can be
used in the VRT. All very much a work in progress...
}
\author{
\strong{Maintainer}: Hugh Graham \email{hugh.graham@permianglobal.com} (\href{https://orcid.org/0000-0001-9451-5010}{ORCID})

}
