% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac-utils.R
\name{vrtility_usr_agent}
\alias{vrtility_usr_agent}
\title{#' User agent for vrtility
#' @return A user agent string
#' @keywords internal
#' @noRd}
\usage{
vrtility_usr_agent()
}
\description{
#' User agent for vrtility
#' @return A user agent string
#' @keywords internal
#' @noRd
}
