% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalraster-tools.R
\name{vector_to_MEM}
\alias{vector_to_MEM}
\title{Create an in-memory GDAL raster dataset from R data}
\usage{
vector_to_MEM(
  data,
  nbands = NULL,
  xsize = NULL,
  ysize = NULL,
  bbox = NULL,
  srs = NULL
)
}
\arguments{
\item{data}{numeric vector, matrix or array}

\item{nbands}{numeric number of bands}

\item{xsize}{numeric number of pixels in x direction}

\item{ysize}{numeric number of pixels in y direction}

\item{bbox}{numeric vector of length 4 representing the bounding box with
the order: xmin, ymin, xmax, ymax}

\item{srs}{character spatial reference system string}
}
\value{
GDALRasterDataset object
}
\description{
Create an in-memory GDAL raster dataset from R data
}
\examples{
v2m <- vector_to_MEM(
  as.vector(c(
    sin(
      sqrt(outer(1:120 - 60, 1:110 - 55, function(x, y) x^2 + y^2)) * 0.9
    )
  )),
  nbands = 1,
  xsize = 120,
  ysize = 110
)

v2m
plot(v2m)
}
