% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai-mgmt.R
\name{n_daemons}
\alias{n_daemons}
\alias{daemons_load_vrtility}
\title{Mirai Daemon Management}
\usage{
n_daemons()

daemons_load_vrtility(py_pkgs = "numpy", ...)
}
\arguments{
\item{py_pkgs}{A character vector of python packages to ensure are loaded
in the mirai daemons.}

\item{...}{Additional arguments passed to \code{\link{vrtility_py_require}}}
}
\description{
Just some helpful functions for managing the mirai daemons and
parallel processing.

Load vrtility python environment in mirai daemons
}
\details{
\code{n_daemons()} returns the number of mirai daemons running.
}
