% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai-mgmt.R
\name{using_daemons}
\alias{using_daemons}
\alias{n_daemons}
\alias{machine_cores}
\title{Mirai Daemon Management}
\usage{
using_daemons()

n_daemons()

machine_cores(all.tests = FALSE, logical = TRUE)
}
\arguments{
\item{all.tests}{Logical: if true apply all known tests.}

\item{logical}{Logical: if possible, use the number of physical CPUs/cores
(if FALSE) or logical CPUs (if TRUE). Currently this is honoured only on
macOS, Solaris and Windows.}
}
\description{
Just some helpful functions for managing the mirai daemons and
parallel processing.
}
\details{
\code{using_daemons()} returns \code{TRUE} if there are any mirai daemons running.

\code{n_daemons()} returns the number of mirai daemons running.

\code{machine_cores()} returns the number of cores on the machine. This function
is from \code{\link[parallel]{detectCores}} and is used detect the number of
cores on the machine.
}
