% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirai-mgmt.R
\name{n_daemons}
\alias{n_daemons}
\title{Mirai Daemon Management}
\usage{
n_daemons()
}
\description{
Just some helpful functions for managing the mirai daemons and
parallel processing.
}
\details{
\code{n_daemons()} returns the number of mirai daemons running.
}
