% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-raster-src.R, R/plot_raster2.R
\name{plot_raster_src}
\alias{plot_raster_src}
\alias{plot.vrt_block}
\alias{plot.vrt_stack}
\alias{plot.vrt_stack_warped}
\alias{plot.vrt_collection}
\alias{plot.Rcpp_GDALRaster}
\title{plot a raster file or \code{vrt_x} object}
\usage{
plot_raster_src(
  x,
  bands = 1,
  col = grDevices::hcl.colors(10, "Viridis"),
  title = c("description", "dttm", "none"),
  rgb_trans = c("linear", "gamma", "hist", "hist_all", "none"),
  col_tbl = NULL,
  maxColorValue = 1,
  normalize = TRUE,
  minmax_def = NULL,
  minmax_pct_cut = NULL,
  xlim = NULL,
  ylim = NULL,
  interpolate = TRUE,
  axes = TRUE,
  main = "",
  xlab = "",
  ylab = "",
  legend = if (length(bands) == 1) TRUE else FALSE,
  digits = 2,
  na_col = grDevices::rgb(0, 0, 0, 0),
  ...
)

\method{plot}{vrt_block}(x, ..., config_options = gdal_config_opts())

\method{plot}{vrt_stack}(x, ..., config_options = gdal_config_opts())

\method{plot}{vrt_stack_warped}(x, ..., config_options = gdal_config_opts())

\method{plot}{vrt_collection}(x, item, ..., config_options = gdal_config_opts())

\method{plot}{Rcpp_GDALRaster}(
  data,
  xsize = NULL,
  ysize = NULL,
  bands = NULL,
  max_pixels = 2.5e+07,
  scale_values = TRUE,
  col_tbl = NULL,
  maxColorValue = 1,
  normalize = TRUE,
  minmax_def = NULL,
  minmax_pct_cut = NULL,
  col = NULL,
  pixel_fn = NULL,
  xlim = NULL,
  ylim = NULL,
  interpolate = TRUE,
  axes = TRUE,
  main = "",
  xlab = "x",
  ylab = "y",
  legend = FALSE,
  digits = 2,
  na_col = grDevices::rgb(0, 0, 0, 0),
  mar = c(0, 0, 0, 0),
  ...
)
}
\arguments{
\item{x}{A path to a raster file}

\item{bands}{a numeric vector of band numbers to plot must be of length
1 or 3.}

\item{col}{Colors to interpolate; must be a valid argument to
\code{\link[grDevices]{col2rgb}}}

\item{title}{a character string indicating what to use as the title of the
plot. One of "description", "dttm", or "none". If "description" the band
description (name) is used, if "dttm" the datetime of the block is used, if
"none" no title is used. Ignored if main is provided.}

\item{col_tbl}{A color table as a matrix or data frame with four or five
columns. Column 1 contains the numeric pixel values. Columns 2:4 contain
the intensities of the red, green and blue primaries (\code{0:1} by default,
or use integer \code{0:255} by setting \code{maxColorValue = 255}).
An optional column 5 may contain alpha transparency values, \code{0} for fully
transparent to \code{1} (or \code{maxColorValue}) for opaque (the default if column 5
is missing). If \code{data} is a \code{GDALRaster} object, a built-in color table will
be used automatically if one exists in the dataset.}

\item{maxColorValue}{A number giving the maximum of the color values range
in \code{col_tbl} (see above). The default is \code{1}.}

\item{normalize}{Logical. \code{TRUE} to rescale pixel values so that their
range is \verb{[0,1]}, normalized to the full range of the pixel data by default
(\code{min(data)}, \code{max(data)}, per band). Ignored if \code{col_tbl} is used.
Set \code{normalize} to \code{FALSE} if a color map function is used that
operates on raw pixel values (see \code{col_map_fn} below).}

\item{minmax_def}{Normalize to user-defined min/max values (in terms of
the pixel data, per band). For single-band grayscale, a numeric vector of
length two containing min, max. For 3-band RGB, a numeric vector of length
six containing b1_min, b2_min, b3_min, b1_max, b2_max, b3_max.}

\item{minmax_pct_cut}{Normalize to a truncated range of the pixel data using
percentile cutoffs (removes outliers). A numeric vector of length two giving
the percentiles to use (e.g., \code{c(2, 98)}). Applied per band. Ignored if
\code{minmax_def} is used.}

\item{xlim}{Numeric vector of length two giving the x coordinate range.
If \code{data} is a \code{GDALRaster} object, the default is the raster xmin, xmax in
georeferenced coordinates, otherwise the default uses pixel/line
coordinates (\code{c(0, xsize)}).}

\item{ylim}{Numeric vector of length two giving the y coordinate range.
If \code{data} is a \code{GDALRaster} object, the default is the raster ymin, ymax in
georeferenced coordinates, otherwise the default uses pixel/line
coordinates (\code{c(ysize, 0)}).}

\item{interpolate}{Logical indicating whether to apply linear interpolation
to the image when drawing (default \code{TRUE}).}

\item{axes}{Logical. \code{TRUE} to draw axes (the default).}

\item{main}{The main title (on top).}

\item{xlab}{Title for the x axis (see \code{?title}).}

\item{ylab}{Title for the y axis (see \code{?title}).}

\item{legend}{Logical indicating whether to include a legend on the plot.
Currently, legends are only supported for continuous data. A color table
will be used if one is specified or the raster has a built-in color table,
otherwise the value for \code{col_map_fn} will be used.}

\item{digits}{The number of digits to display after the decimal point in
the legend labels when raster data are floating point.}

\item{na_col}{Color to use for \code{NA} as a 7- or 9-character hexadecimal code.
The default is transparent (\code{"#00000000"}, the return value of
\code{rgb(0,0,0,0)}).}

\item{...}{Other parameters to be passed to \code{plot.default()}.}

\item{config_options}{A named character vector of gdal config options to set
before attempting to read the VRT.}

\item{item}{The numeric index of the item, in the vrt_collection, to plot}

\item{data}{Either a \code{GDALRaster} object from which data will be read, or
a numeric vector of pixel values arranged in left to right, top to
bottom order, or a list of band vectors. If input is vector or list,
the information in attribute \code{gis} will be used if present (see \code{\link[gdalraster:read_ds]{read_ds()}}),
potentially ignoring values below for \code{xsize}, \code{ysize}, \code{nbands}.}

\item{xsize}{The number of pixels along the x dimension in \code{data}. If \code{data}
is a \code{GDALRaster} object, specifies the size at which the raster will be
read (used for argument \code{out_xsize} in \code{GDALRaster$read()}). By default,
the entire raster will be read at full resolution.}

\item{ysize}{The number of pixels along the y dimension in \code{data}. If \code{data}
is a \code{GDALRaster} object, specifies the size at which the raster will be
read (used for argument \code{out_ysize} in \code{GDALRaster$read()}). By default,
the entire raster will be read at full resolution.}

\item{max_pixels}{The maximum number of pixels that the function will
attempt to display (per band). An error is raised if \code{(xsize * ysize)}
exceeds this value. Setting to \code{NULL} turns off this check.}

\item{pixel_fn}{An optional function that will be applied to the input
pixel data. Must accept vector input and return a numeric vector of the same
length as its input.}

\item{pal}{a character vector of colors to use when plotting a single band.
for example when sea presents as very dark.}
}
\description{
A wrapper around \code{gdalraster::plot_raster} that simplifies the
process of plotting a raster file.
}
