% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-utils.R
\name{bbox_to_projected}
\alias{bbox_to_projected}
\alias{ogr_bbox_from_file}
\alias{ogr_srs_from_file}
\title{Project a lat/long bounding box to a generic projected coordinate system}
\usage{
bbox_to_projected(
  x,
  proj_specific = NULL,
  proj_generic = c("laea", "aeqd", "utm", "pconic", "eqdc"),
  ellps = "WGS84",
  no_defs = TRUE,
  opts = ""
)

ogr_bbox_from_file(x, latlon = FALSE, extend_x = NULL, extend_y = NULL)

ogr_srs_from_file(x)
}
\arguments{
\item{x}{character vector. Path to a spatial vector file.}

\item{proj_specific}{a character vector. The projection to use.
A PROJ-readable string or an EPSG code. If NULL, a generic projection
will be used.}

\item{proj_generic}{a character vector. The projection to use. One of
"laea", "aeqd", "utm", "pconic", or "eqdc".}

\item{ellps}{a character vector. The ellipsoid to use. Select from
\code{sf_proj_info(type = "ellps")}.}

\item{no_defs}{a logical. Whether to include the +no_defs option in the proj
string.}

\item{opts}{a character vector. Additional proj options to pass to the
proj string. see details for more information.}

\item{latlon}{logical. Whether to return the bounding box in lat/long
(EPSG:4326). Default is FALSE.}

\item{extend_x}{numeric. Amount to extend the bounding box in the x
direction (both min and max).}

\item{extend_y}{numeric. Amount to extend the bounding box in the y
direction (both min and max).}
}
\value{
a numeric vector of length 4 representing the projected bounding box
in the new coordinate system. Attributes include the new proj4 and wkt
string.

A numeric vector of length 4 representing the bounding box ordered
as: xmin, ymin, xmax, ymax.

A character string representing the SRS in WKT format.
}
\description{
This function takes a lat/long bounding box and projects it to
either a prescibed projected coordinate system or a generic projected
coordinate system suitable for the AOI.
}
\details{
For further info about the available "generic" projects see:
for utm: \url{https://proj.org/en/9.4/operations/projections/utm.html}
for laea: \url{https://proj.org/en/9.4/operations/projections/laea.html}
for aeqd: \url{https://proj.org/en/9.4/operations/projections/aeqd.html}
for pconic: \url{https://proj.org/en/9.4/operations/projections/pconic.html}
for eqdc: \url{https://proj.org/en/9.4/operations/projections/eqdc.html}
}
\examples{
bbox <- gdalraster::bbox_from_wkt(
  wkt = "POINT (144.3 -7.6)",
  extend_x = 0.17,
  extend_y = 0.125
)

bbox_to_projected(bbox)

bbox_to_projected(bbox, proj_generic = "utm")


\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
bbox_to_projected(
 c(-3.56, 50.69, -3.46, 50.75),
 proj_specific = "EPSG:27700"
)
\dontshow{\}) # examplesIf}
}
