% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stac-utils.R
\name{stac_query}
\alias{stac_query}
\alias{sign_planetary_computer}
\alias{sentinel2_stac_query}
\title{Query a STAC source}
\usage{
stac_query(
  bbox,
  stac_source,
  collection,
  start_date,
  end_date,
  limit = 999,
  ...
)

sign_planetary_computer(items, subscription_key = Sys.getenv("rsi_pc_key"))

sentinel2_stac_query(
  bbox,
  start_date,
  end_date,
  assets = c("B01", "B02", "B03", "B04", "B05", "B06", "B07", "B08", "B8A", "B09", "B11",
    "B12", "SCL"),
  max_cloud_cover = 10,
  stac_source = "https://planetarycomputer.microsoft.com/api/stac/v1/",
  collection = "sentinel-2-l2a",
  mpc_sign = TRUE
)
}
\arguments{
\item{bbox}{A numeric vector of the bounding box (length 4) in lat/long}

\item{stac_source}{A character string of the STAC source}

\item{collection}{A character string of the collection to query}

\item{start_date}{A character string of the start date}

\item{end_date}{A character string of the end date}

\item{limit}{The number of items to return}

\item{...}{Additional arguments to pass to the query}

\item{items}{A STACItemCollection.}

\item{subscription_key}{Optionally, a subscription key associated with your
Planetary Computer account. At the time of writing, this is required for
downloading Sentinel 1 RTC products, as well as NAIP imagery. This key will
be automatically used if the environment variable \code{rsi_pc_key} is set.}

\item{assets}{A character vector of the asset names to include}

\item{max_cloud_cover}{A numeric value of the maximum cloud cover percentage}

\item{mpc_sign}{A logical indicating whether to sign the items using the
Planetary Computer API's signing method (only required if using the
Planetary Computer STAC API).}
}
\value{
A list of items

A STACItemCollection object with signed assets url.

A stac_vrt object
}
\description{
Query a STAC source

Sign STAC items retrieved from the Planetary Computer

Generate a Sentinel 2 stac collection doc_imes object
}
\details{
copied from the \code{rsi} package
}
