% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-add-empty-band.R
\name{vrt_add_empty_band}
\alias{vrt_add_empty_band}
\title{Add an empty band to a VRT_x object}
\usage{
vrt_add_empty_band(x, after, description, save_dir)
}
\arguments{
\item{x}{A VRT_x object}

\item{after}{numeric indicating the band number to add the empty band after}
}
\description{
Add an empty band to a VRT_x object
}
