% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-classes.R, R/vrt-collect.R, R/vrt-stack.R
\name{print.vrt_stack}
\alias{print.vrt_stack}
\alias{print.vrt_collection}
\alias{save_vrt}
\alias{save_vrt.vrt_stack}
\alias{vrt_collect}
\alias{vrt_collect.doc_items}
\alias{vrt_stack}
\alias{vrt_stack.vrt_collection}
\title{Print a vrt_block object}
\usage{
\method{print}{vrt_stack}(x, xml = FALSE, pixfun = FALSE, maskfun = FALSE, ...)

\method{print}{vrt_collection}(x, xml = FALSE, pixfun = FALSE, maskfun = FALSE, blocks = FALSE, ...)

save_vrt(x, outfile)

\method{save_vrt}{vrt_stack}(
  x,
  outfile = fs::file_temp(tmp_dir = getOption("vrt.cache"), ext = "vrt")
)

vrt_collect(x)

\method{vrt_collect}{doc_items}(x)

vrt_stack(x, ...)

\method{vrt_stack}{vrt_collection}(x, quiet = TRUE, ...)
}
\arguments{
\item{x}{A vrt_collection object}

\item{xml}{A logical indicating whether to print the XML}

\item{pixfun}{A logical indicating whether to print the pixel function}

\item{maskfun}{A logical indicating whether to print the mask function}

\item{...}{Additional arguments passed to methods}

\item{blocks}{A logical indicating whether to print the blocks instead of
the collection summary.}

\item{outfile}{A character string of the output file}

\item{quiet}{Logical. If TRUE, suppress GDAL progress bar}
}
\description{
Print a vrt_block object

Save a vrt_block object to disk

Construct the base VRT object for composing VRT pipelines.

Stack VRT files from a vrt_collection object
}
