% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-collect.R
\name{vrt_collect}
\alias{vrt_collect}
\alias{vrt_collect.doc_items}
\alias{print.vrt_collection}
\title{Construct the base VRT object for composing VRT pipelines.}
\usage{
vrt_collect(
  x,
  t_srs,
  te,
  tr,
  mask_band = NULL,
  resampling = c("blinear", "near", "cubic", "cubicspline", "lanczos", "average", "rms",
    "mode", "max", "min", "med", "q1", "q3", "sum"),
  quiet = TRUE
)

\method{vrt_collect}{doc_items}(
  x,
  t_srs,
  te,
  tr,
  mask_band = NULL,
  resampling = c("bilinear", "near", "cubic", "cubicspline", "lanczos", "average", "rms",
    "mode", "max", "min", "med", "q1", "q3", "sum"),
  quiet = TRUE
)

\method{print}{vrt_collection}(x, xml = FALSE, pixfun = FALSE, maskfun = FALSE, blocks = FALSE, ...)
}
\arguments{
\item{x}{An object to be used to create a vrt_x object see details.}

\item{t_srs}{character target SRS must be a numeric EPSG code, or SRS like
character such as a proj4 string or WKT.}

\item{te}{numeric vector of the target extent in the form
c(xmin, ymin, xmax, ymax) using the same SRS as in \code{t_srs}.}

\item{tr}{numeric vector of the target resolution in the form c(xres, yres)}

\item{quiet}{logical indicating whether to suppress progress bar.}

\item{blocks}{A logical indicating whether to print the blocks instead of
the collection summary.}

\item{...}{Additional arguments not used}
}
\description{
Construct the base VRT object for composing VRT pipelines.
}
\details{
For now the main way to create a vrt_collection object, which forms
the basis of the vrrt-based pipelines in vrtility is using a doc_items
object from the \code{rstac} package. For more info on how to create a doc_items
object see \code{\link[=sentinel2_stac_query]{sentinel2_stac_query()}}. To build a vrt_stack object a vrt_collection
is required first. The vrt_collection object is essentially a list of warped
VRT files. When vrt_stack is called on a vrt_collection object, the VRT files
are stacked into a single VRT file where each band in the VRT contains the
files of multiple epochs. We can then composite or summarise these layers
using the pixel functions.
}
