% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-compute.R
\name{vrt_compute}
\alias{vrt_compute}
\alias{vrt_compute.vrt_block}
\alias{vrt_compute.vrt_stack_warped}
\alias{vrt_compute.vrt_stack}
\alias{vrt_compute.vrt_collection}
\title{Generate a composite raster from (virtual) raster sources.}
\usage{
vrt_compute(
  x,
  outfile,
  t_srs,
  te,
  tr,
  warp_options = getOption("vrt.gdal.warp.options"),
  config_options = getOption("vrt.gdal.config.options"),
  quiet = FALSE
)

\method{vrt_compute}{vrt_block}(
  x,
  outfile,
  t_srs = x$srs,
  te = x$bbox,
  tr = x$res,
  warp_options = getOption("vrt.gdal.warp.options"),
  config_options = getOption("vrt.gdal.config.options"),
  quiet = FALSE
)

\method{vrt_compute}{vrt_stack_warped}(
  x,
  outfile,
  t_srs = x$srs,
  te = x$bbox,
  tr = x$res,
  warp_options = getOption("vrt.gdal.warp.options"),
  config_options = getOption("vrt.gdal.config.options"),
  quiet = FALSE
)

\method{vrt_compute}{vrt_stack}(
  x,
  outfile,
  t_srs,
  te,
  tr,
  warp_options = getOption("vrt.gdal.warp.options"),
  config_options = getOption("vrt.gdal.config.options"),
  quiet = FALSE
)

\method{vrt_compute}{vrt_collection}(
  x,
  outfile,
  t_srs = x$srs,
  te = x$bbox,
  tr = x$res,
  warp_options = getOption("vrt.gdal.warp.options"),
  config_options = getOption("vrt.gdal.config.options"),
  quiet = FALSE
)
}
\arguments{
\item{x}{A vrt_block, vrt_stack, or vrt_collection object}

\item{outfile}{A character string of the output file path}

\item{t_srs}{A character string of the target SRS}

\item{te}{A numeric vector of the target extent in the form
c(xmin, ymin, xmax, ymax) and must be the same SRS as in \code{t_srs}.}

\item{tr}{A numeric vector of the target resolution in the form c(xres, yres)}

\item{warp_options}{A character vector of options to pass to the warp}

\item{config_options}{A character vector of options to set in the GDAL
environment}

\item{quiet}{A logical indicating whether to suppress output}
}
\value{
A character string of the path to the output raster
}
\description{
Generate a composite raster from (virtual) raster sources.
}
\details{
This is the primary function to call processing of raster data. The
behaviour of the warper is dependent on the form of the input vrt datasets
and the associated options.
}
