% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-derived-block.R
\name{vrt_derived_block}
\alias{vrt_derived_block}
\title{Create vrt blocks with derived bands.}
\usage{
vrt_derived_block(x, ...)
}
\arguments{
\item{x}{a \code{vrt_block} or \code{vrt_collection} object.}

\item{...}{named R formulas for new bands to create. Formula variable names
must the names of existing bands in the vrt_block object. the argument name
is used to name the new (derived) band.}
}
\description{
Create new vrt blocks containing derived expression bands.
}
\examples{
\dontshow{if (check_muparser()) withAutoprint(\{ # examplesIf}
s2files <- fs::dir_ls(system.file("s2-data", package = "vrtility"))

ex_collect <- vrt_collect(s2files[1])

ex_ndvi <- vrt_derived_block(
  ex_collect,
  ndvi ~ (B08 - B04) / (B08 + B04)
)
plot(ex_ndvi)
\dontshow{\}) # examplesIf}
}
