% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-move-band.R
\name{vrt_move_band}
\alias{vrt_move_band}
\title{Add an empty band to a VRT_x object}
\usage{
vrt_move_band(x, band_idx, after, save_dir)
}
\arguments{
\item{x}{A VRT_x object}

\item{after}{numeric indicating the band number to add the empty band after}
}
\description{
Add an empty band to a VRT_x object
}
