% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt_save.R
\name{vrt_save}
\alias{vrt_save}
\title{Save a vrt_block object to disk}
\usage{
vrt_save(x, outfile)
}
\arguments{
\item{x}{A \code{vrt_stack} of \code{vrt_block} object.}

\item{outfile}{A character string of the output file}
}
\description{
Save a vrt_block object to disk
}
