% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-set-gdal-pixfun.R
\name{vrt_set_gdal_pixelfun}
\alias{vrt_set_gdal_pixelfun}
\alias{vrt_set_gdal_pixelfun.vrt_block}
\alias{vrt_set_gdal_pixelfun.vrt_collection}
\title{Set built-in GDAL pixel functions of a VRT stack object}
\usage{
vrt_set_gdal_pixelfun(x, pixfun, ..., band_idx)

\method{vrt_set_gdal_pixelfun}{vrt_block}(x, pixfun, ..., band_idx = NULL)

\method{vrt_set_gdal_pixelfun}{vrt_collection}(x, pixfun, ..., band_idx = NULL)
}
\arguments{
\item{x}{A vrt_stack object}

\item{pixfun}{A built-in GDAL pixel function. See details for a list of
available functions.}

\item{...}{named arguments used within the pixel function see details.}

\item{band_idx}{The indices of the bands to set the pixel function for. If
NULL, the pixel function is set for all bands.}
}
\description{
Set built-in GDAL pixel functions of a VRT stack object
}
\details{
The documentation for the GDAL pixel functions can be found at
\url{https://gdal.org/en/stable/drivers/raster/vrt.html#built-in-pixel-functions}

Where a pixel function requires arguments, these can be passed as named
arguments, in line with the specified naming in PixelFunctionArguments of
the table at the above link.
}
\examples{
s2files <- fs::dir_ls(system.file("s2-data", package = "vrtility"))
ex_collect <- vrt_collect(s2files)

t_block <- ex_collect[[1]][[1]]
ex_stack <- ex_collect |>
  vrt_set_maskfun(
    mask_band = "SCL",
    mask_values = c(0, 1, 2, 3, 8, 9, 10, 11),
    drop_mask_band = TRUE
  ) |>
  vrt_warp(
    t_srs = t_block$srs,
    te = t_block$bbox,
    tr = t_block$res
  ) |>
  vrt_stack()

plot(
  vrt_set_gdal_pixelfun(
    ex_stack,
    pixfun = "min",
    propagateNoData = TRUE
  ),
  bands = c(3, 2, 1)
)

plot(
  vrt_set_gdal_pixelfun(
    ex_stack,
    pixfun = "min"
  ),
  bands = c(3, 2, 1)
)

}
