% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt_set_mask.R
\name{vrt_set_mask}
\alias{vrt_set_mask}
\alias{vrt_set_mask.vrt_block}
\title{Set mask band of a VRT collection}
\usage{
vrt_set_mask(x, ...)

\method{vrt_set_mask}{vrt_block}(x, mask_band)
}
\description{
Set mask band of a VRT collection
}
