% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask-pixel-funs.R, R/vrt-set-mask.R
\name{set_mask_numpy}
\alias{set_mask_numpy}
\alias{build_intmask}
\alias{build_bitmask}
\alias{vrt_set_maskfun}
\alias{vrt_set_maskfun.vrt_block}
\alias{vrt_set_maskfun.vrt_collection}
\title{Masking functions for pixel-based processing}
\usage{
set_mask_numpy()

build_intmask()

build_bitmask()

vrt_set_maskfun(
  x,
  mask_band,
  mask_values,
  build_mask_pixfun,
  set_mask_pixfun,
  drop_mask_band
)

\method{vrt_set_maskfun}{vrt_block}(
  x,
  mask_band,
  mask_values,
  build_mask_pixfun = vrtility::build_intmask(),
  set_mask_pixfun = vrtility::set_mask_numpy(),
  drop_mask_band = TRUE
)

\method{vrt_set_maskfun}{vrt_collection}(
  x,
  mask_band,
  mask_values,
  build_mask_pixfun = vrtility::build_intmask(),
  set_mask_pixfun = vrtility::set_mask_numpy(),
  drop_mask_band = TRUE
)
}
\arguments{
\item{x}{A VRT collection}

\item{mask_band}{The name of the mask band}

\item{mask_values}{A numeric vector of integer or bit values to be masked.}

\item{build_mask_pixfun}{A character string of the Python code to build the
mask. Provided functions include \code{\link[=build_intmask]{build_intmask()}} and \code{\link[=build_bitmask]{build_bitmask()}}.
See details.}

\item{set_mask_pixfun}{A character string of the Python code to set the mask.
Provided functions include \code{\link[=set_mask_numpy]{set_mask_numpy()}}. See
details.}

\item{drop_mask_band}{Logical. If TRUE, the mask band will be removed from
the VRT block.}
}
\value{
A VRT block with the mask band set.

A VRT collection with the mask band set.
}
\description{
Masking functions for pixel-based processing

Masking functions for pixel-based processing

Masking functions for pixel-based processing

Set mask band of a VRT collection
}
\details{
This function is a simple bit mask function that can be used to
mask out pixels based on a bit mask. There is no bitwise transformation
applied to the provided valid_bits.

This function is a simple integer mask function that can be used to
mask out pixels based on a bit mask. There is no bitwise transformation
applied to the provided valid_bits.

This function is a simple bit-wise mask function that can be used to
mask out pixels based on a bit mask. The valid_bits are used to determine
which bits to mask out.

The \code{build_mask_pixfun} function is used to build the mask band. Where the
mask band is a true bitmask and bit-wise operations are required, the
\code{\link[=build_bitmask]{build_bitmask()}} function should be used. For integer-based masking, where
the mask band is provided as a single band with integer values, the
\code{\link[=build_intmask]{build_intmask()}} function should be used.

The \code{set_mask_pixfun} function is used to apply the mask to the other bands.
In general \code{\link[=set_mask_numpy]{set_mask_numpy()}} should be used, although a custom function can
be provided.
}
