% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask-pixel-funs.R, R/vrt-set-mask.R
\name{bitmask_numpy}
\alias{bitmask_numpy}
\alias{bitmask_numba}
\alias{build_bitmask_numpy}
\alias{vrt_set_maskfun}
\alias{vrt_set_maskfun.vrt_collection}
\title{Masking functions for pixel-based processing}
\usage{
bitmask_numpy()

bitmask_numba()

build_bitmask_numpy()

vrt_set_maskfun(
  x,
  valid_bits,
  mask_pix_fun = vrtility::bitmask_numba,
  drop_mask_band = TRUE,
  ...
)

\method{vrt_set_maskfun}{vrt_collection}(
  x,
  valid_bits,
  mask_pix_fun = vrtility::bitmask_numba,
  drop_mask_band = TRUE
)
}
\arguments{
\item{x}{A VRT collection}
}
\value{
A VRT collection with the mask band set.
}
\description{
Masking functions for pixel-based processing

Masking functions for pixel-based processing

Masking functions for pixel-based processing

Set mask band of a VRT collection
}
\details{
This function is a simple bit mask function that can be used to
mask out pixels based on a bit mask. There is no bitwise transformation
applied to the provided valid_bits.

This function is a simple bit mask function that can be used to
mask out pixels based on a bit mask. There is no bitwise transformation
applied to the provided valid_bits.

This function is a simple bit mask function that can be used to
mask out pixels based on a bit mask. There is no bitwise transformation
applied to the provided valid_bits.
}
