% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-set-mask.R
\name{vrt_set_maskfun}
\alias{vrt_set_maskfun}
\alias{vrt_set_maskfun.vrt_block}
\alias{vrt_set_maskfun.vrt_collection}
\title{Set mask band of a VRT collection}
\usage{
vrt_set_maskfun(
  x,
  mask_band,
  mask_values,
  build_mask_pixfun,
  set_mask_pixfun,
  drop_mask_band,
  cache_dir
)

\method{vrt_set_maskfun}{vrt_block}(
  x,
  mask_band,
  mask_values,
  build_mask_pixfun = vrtility::build_intmask(),
  set_mask_pixfun = vrtility::set_mask_numpy(),
  drop_mask_band = TRUE,
  cache_dir = getOption("vrt.cache")
)

\method{vrt_set_maskfun}{vrt_collection}(
  x,
  mask_band,
  mask_values,
  build_mask_pixfun = vrtility::build_intmask(),
  set_mask_pixfun = vrtility::set_mask_numpy(),
  drop_mask_band = TRUE,
  cache_dir = getOption("vrt.cache")
)
}
\arguments{
\item{x}{A VRT collection}

\item{mask_band}{The name of the mask band}

\item{mask_values}{A numeric vector of integer or bit values to be masked.}

\item{build_mask_pixfun}{A character string of the Python code to build the
mask. Provided functions include \code{\link[=build_intmask]{build_intmask()}} and \code{\link[=build_bitmask]{build_bitmask()}}.
See details.}

\item{set_mask_pixfun}{A character string of the Python code to set the mask.
Provided functions include \code{\link[=set_mask_numpy]{set_mask_numpy()}}. See
details.}

\item{drop_mask_band}{Logical. If TRUE, the mask band will be removed from
the VRT block.}

\item{cache_dir}{A character string of the directory to use for temporary
files. In general this should be left alone. main purpose is to manage cache
location when running asyncronously with mirai.}
}
\value{
A VRT block with the mask band set.

A VRT collection with the mask band set.
}
\description{
Set mask band of a VRT collection
}
\details{
The \code{build_mask_pixfun} function is used to build the mask band. Where the
mask band is a true bitmask and bit-wise operations are required, the
\code{\link[=build_bitmask]{build_bitmask()}} function should be used. For integer-based masking, where
the mask band is provided as a single band with integer values, the
\code{\link[=build_intmask]{build_intmask()}} function should be used.

The \code{set_mask_pixfun} function is used to apply the mask to the other bands.
In general \code{\link[=set_mask_numpy]{set_mask_numpy()}} should be used, although a custom function can
be provided.
}
