% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-set-pixfun.R
\name{vrt_set_pixelfun}
\alias{vrt_set_pixelfun}
\alias{vrt_set_pixelfun.vrt_stack}
\title{Set the pixel function of a VRT stack object}
\usage{
vrt_set_pixelfun(x, pixfun)

\method{vrt_set_pixelfun}{vrt_stack}(x, pixfun = vrtility::median_numpy())
}
\arguments{
\item{x}{A vrt_stack object}

\item{pixfun}{A function that returns the Python code for the pixel function}
}
\description{
Set the pixel function of a VRT stack object
}
