% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-set-pixfun.R, R/zvrt-pixel-funs-composite.R
\name{vrt_set_py_pixelfun}
\alias{vrt_set_py_pixelfun}
\alias{vrt_set_py_pixelfun.vrt_stack}
\alias{median_numpy}
\alias{mean_numpy}
\alias{geomean_numpy}
\alias{mean_db_numpy}
\title{Set the pixel function of a VRT stack object}
\usage{
vrt_set_py_pixelfun(x, pixfun)

\method{vrt_set_py_pixelfun}{vrt_stack}(x, pixfun = vrtility::median_numpy())

median_numpy()

mean_numpy()

geomean_numpy()

mean_db_numpy()
}
\arguments{
\item{x}{A vrt_stack object}

\item{pixfun}{A function that returns the Python code for the pixel function}
}
\value{
character of the python function

character of the python function
}
\description{
Set the pixel function of a VRT stack object
}
\details{
\code{median_numpy} is a pixel function that calculates the median of
the input arrays, it is injected into the VRT file as a Python function.
\code{mean_numpy} works in the same way but calculates the mean.`

\code{geomean_numpy} is a pixel function that calculates the geometric
mean of the input arrays. Use cases of this are at present unclear to me.
If you have thoughts or references please let me know.

\code{mean_db_numpy} is a pixel function that calculates the mean of
the input arrays and then converts to decibels. This is useful for
calculating the mean of radar raw/linear backscatter values, for example.
}
