% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-set-scale.R
\name{vrt_set_scale}
\alias{vrt_set_scale}
\title{set the scale values for a VRT_x object}
\usage{
vrt_set_scale(x, scale_value, offset_value, band_idx)
}
\arguments{
\item{x}{A VRT_x object}

\item{scale_value}{a numeric vector of length 1 or equal to the number of
bands in the VRT_x object.}

\item{offset_value}{a numeric value to set the Offset XML tag for the VRT_x
object. Default is 0.}

\item{band_idx}{numeric; the target band position(s) to set the scale value
for. If NULL, The scale_value will be set for all bands.}
}
\value{
A modified object of the same class as \code{x} with scale and offset
values set.
}
\description{
Set the Scale XML tag for a VRT_x object. This information is
preserved in derived VRT and GeoTIFF files.
}
\details{
This sets the Scale and Offset XML tags for the VRT_x object. This is
useful for scaling the pixel values of the VRT_x object. This is however,
a little confusing - the actual values are not altered but the scaling value
is applied when read by most software.

To unset a scale or offset simply set scale= 1 and offset = 0.
}
\examples{
s2files <- fs::dir_ls(system.file("s2-data", package = "vrtility"))
ex_collect <- vrt_collect(s2files)
ex_sc1 <- vrt_set_scale(
  ex_collect,
  scale_value = 1e-4
)
print(ex_sc1[[1]][[1]], xml = TRUE)
}
