% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-stack.R
\name{vrt_stack}
\alias{vrt_stack}
\alias{vrt_stack.vrt_collection}
\alias{print.vrt_stack}
\title{Stack VRT files from a vrt_collection object}
\usage{
vrt_stack(x)

\method{vrt_stack}{vrt_collection}(x)

\method{print}{vrt_stack}(x, xml = FALSE, pixfun = FALSE, maskfun = FALSE, ...)
}
\arguments{
\item{x}{A vrt_collection object}

\item{xml}{A logical indicating whether to print the XML}

\item{pixfun}{A logical indicating whether to print the pixel function}

\item{maskfun}{A logical indicating whether to print the mask function}

\item{...}{Additional arguments not used}
}
\value{
A vrt_stack object
}
\description{
Stack VRT files from a vrt_collection object

Print a vrt_block object
}
\details{
This function stacks VRT files from a vrt_collection object into a
single VRT file containing multiple layers for each RasterBand. The VRT
files are stacked in the order they are provided in the vrt_collection
object. If this is derived from a rstac object, the order should be ordered
by date.
}
