% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-stack.R
\name{vrt_stack}
\alias{vrt_stack}
\alias{vrt_stack.vrt_collection}
\alias{print.vrt_stack}
\alias{save_vrt}
\alias{save_vrt.vrt_block}
\title{Stack VRT files from a vrt_collection object}
\usage{
vrt_stack(x, ...)

\method{vrt_stack}{vrt_collection}(x, quiet = TRUE, ...)

\method{print}{vrt_stack}(x, xml = FALSE, pixfun = FALSE, maskfun = FALSE, ...)

save_vrt(x, outfile)

\method{save_vrt}{vrt_block}(
  x,
  outfile = fs::file_temp(tmp_dir = getOption("vrt.cache"), ext = "vrt")
)
}
\arguments{
\item{...}{Additional arguments not used}

\item{quiet}{Logical. If TRUE, suppress GDAL progress bar}

\item{xml}{A logical indicating whether to print the XML}

\item{pixfun}{A logical indicating whether to print the pixel function}

\item{maskfun}{A logical indicating whether to print the mask function}

\item{outfile}{A character string of the output file}
}
\description{
Stack VRT files from a vrt_collection object

Print a vrt_block object

Save a vrt_block object to disk
}
