% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt-warp.R
\name{vrt_warp}
\alias{vrt_warp}
\alias{vrt_warp.vrt_block}
\alias{vrt_warp.vrt_collection}
\title{Generate a composite raster from (virtual) raster sources.}
\usage{
vrt_warp(
  x,
  outfile,
  t_srs,
  te,
  res = 10,
  warp_options = getOption("vrt.gdal.warp.options"),
  config_options = getOption("vrt.gdal.config.options"),
  quiet = FALSE,
  ...
)

\method{vrt_warp}{vrt_block}(
  x,
  outfile,
  t_srs,
  te,
  res = x$res,
  warp_options = getOption("vrt.gdal.warp.options"),
  config_options = getOption("vrt.gdal.config.options"),
  quiet = FALSE
)

\method{vrt_warp}{vrt_collection}(
  x,
  outfile,
  t_srs,
  te,
  res,
  warp_options = getOption("vrt.gdal.warp.options"),
  config_options = getOption("vrt.gdal.config.options"),
  quiet = FALSE
)
}
\arguments{
\item{outfile}{A character string of the output file path}

\item{t_srs}{A character string of the target SRS}

\item{res}{A numeric of the target resolution in units of t_srs}

\item{warp_options}{A character vector of options to pass to the warp}

\item{config_options}{A character vector of options to set in the GDAL
environment}

\item{quiet}{A logical indicating whether to suppress output}

\item{src_files}{A stac_vrt object or a character vector of file paths to the
rasters}

\item{bbox}{A numeric vector of the bounding box (length 4) in lat/long}

\item{fun}{A character string of the pixel function to apply (only median is
supported for now...)}

\item{vrt_options}{A list of options to pass to the VRT}
}
\value{
A character string of the path to the output raster
}
\description{
Generate a composite raster from (virtual) raster sources.
}
\details{
This is the main package function at present. It is quite targetted
towards Sentinel 2 data and will likely change - but this will server as the
main testing point for now!
}
