% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-utils.R
\name{to_wkt}
\alias{to_wkt}
\alias{to_wkt.default}
\alias{to_wkt.numeric}
\alias{to_wkt.character}
\title{Convert an object to wkt.}
\usage{
to_wkt(x)

\method{to_wkt}{default}(x)

\method{to_wkt}{numeric}(x)

\method{to_wkt}{character}(x)
}
\arguments{
\item{x}{The object to convert to wkt.}
}
\value{
A character string of the object in wkt format.
}
\description{
Convert an object to wkt.
}
\keyword{internal}
