% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiband_reduce.R
\name{extract_band_matrices}
\alias{extract_band_matrices}
\alias{restructure_cells}
\title{Extract band matrices using C++}
\usage{
extract_band_matrices(
  x,
  row_indices,
  col_indices,
  n_cells,
  n_timepoints,
  n_bands
)

restructure_cells(cell_vals)
}
\arguments{
\item{x}{A list of lists of matrices, where each inner list represents a
time point - the product of \code{read_block_arrays()}.}

\item{row_indices}{A vector of row indices.}

\item{col_indices}{A vector of column indices.}

\item{n_cells}{The number of cells.}

\item{n_timepoints}{The number of time points.}

\item{n_bands}{The number of bands.}

\item{cell_vals}{A list of lists of vectors, where each inner list represents
a time point - the product of \code{geomedian_reduction()}.}
}
\value{
A list of matrices, where each matrix corresponds to a band.

A numeric vector of length \code{n_bands * n_cells}, where \code{n_bands} is
the number of bands and \code{n_cells} is the number of cells.
}
\description{
Internal C++ function wrapped for R

Internal C++ function wrapped for R
}
\keyword{internal}
